Initial Commit
Project is being remade with much better design principles, as well as all tools integrated in one. The first project was a quick prototype that was very difficult to keep working on.
Some code will be reused, while most will be redesigned from the ground up. Much more functionality will be possible, and it will be easier to implement it for others of various skill levels.

This initial commit is still a WIP.